Moved classes to more specific package
