+++ java/code/src/com/suse/manager/webui/utils/test/InputValidatorTest.java
@@ -17,7 +17,7 @@ package com.suse.manager.webui.utils.test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.suse.manager.webui.controllers.MinionsAPI;
-import com.suse.manager.webui.controllers.utils.RegularMinionBootstrapper;
+import com.suse.manager.webui.controllers.bootstrap.RegularMinionBootstrapper;
 import com.suse.manager.webui.services.iface.SaltApi;
 import com.suse.manager.webui.services.iface.SystemQuery;
 import com.suse.manager.webui.services.test.TestSaltApi;
